Store constant tables in PROGMEM on AVR
When compiling for an Arduino Uno, this saves 4546 bytes of RAM (5860 -> 1214)
at the expense of 302 additional bytes of code. Most of this additional
code is used by the AES implementation, probably because the use of
special PROGMEM loading instructions reduces the optimization's freedom
with regard to instruction ordering and register assignment.