Move DECL_ON_LMIC_EVENT from lmic.c to lmic.h
This declares the onLmicEvent() callback function that must be defined by an
application. Previously, the application was expected to define this
itself, and a reference to it was declared in lmic.c.

By moving the declaration into lmic.h, this prevents problems when an
application uses a different function signature for the function, or
when defining the callback in a C++ file (since lmic.h uses extern "C",
the compiler now knows to generate a C-style function rather than C++).