WIP: Use macros for all constant table definitions and lookups
This should not change any behaviour right now, but makes it easier to
introduce custom constant lookup code in a single spot (e.g. to put
tables in PROGMEM on AVR). To prevent accidental direct access of the
tables, their name is mangled by the macros. To prevent using the wrong
accessor macro (for another array element type), the accessor macros use
an inline function to actually return the values.

Note that the generated code is not totally identical, since the inline
functions seem to cause gcc to take different optimization decisions. On
AVR, this commit actually shrinks the compiled program by hundred bytes
or so.

WIP: This only adds the table macros, but does not use them yet